build 17 - misc

+++ paintball2/qcommon/common.c
@@ -1286,7 +1286,7 @@ void *Z_TagMalloc (int size, int tag)
 Z_Malloc
 ========================
 */
-void *Z_Malloc (int size)
+void *Z_Malloc (size_t size)
 {
 	return Z_TagMalloc(size, 0);
 }
